Use new geometry interface
