+++ L1Trigger/L1THGCal/interface/fe_codecs/HGCalBestChoiceCodec.h
@@ -23,12 +23,12 @@ class HGCalBestChoiceCodec : public HGCalTriggerFE::Codec<HGCalBestChoiceCodec,H
 
         HGCalBestChoiceCodec(const edm::ParameterSet& conf);
 
-        void setDataPayloadImpl(const Module& mod, 
+        void setDataPayloadImpl(const HGCalTriggerGeometryBase& geom, 
                 const HGCEEDigiCollection& ee,
                 const HGCHEDigiCollection& fh,
                 const HGCHEDigiCollection& bh );
 
-        void setDataPayloadImpl(const Module& mod, 
+        void setDataPayloadImpl(const HGCalTriggerGeometryBase& geom, 
                 const l1t::HGCFETriggerDigi& digi);
 
         std::vector<bool> encodeImpl(const data_type&) const ;
