+++ L1Trigger/L1THGCal/interface/fe_codecs/HGCalBestChoiceCodecImpl.h
@@ -4,7 +4,6 @@
 
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "L1Trigger/L1THGCal/interface/HGCalTriggerGeometryBase.h"
-#include "L1Trigger/L1THGCal/interface/HGCalTriggerLightweightGeometryBase.h"
 #include "DataFormats/ForwardDetId/interface/HGCTriggerHexDetId.h"
 #include "DataFormats/HGCDigi/interface/HGCDigiCollections.h"
 
@@ -36,17 +35,16 @@ class HGCalBestChoiceCodecImpl
         std::vector<bool> encode(const data_type&) const ;
         data_type         decode(const std::vector<bool>&) const;  
 
-        void linearize(const HGCalTriggerGeometry::Module& ,
-                const std::vector<HGCDataFrame<HGCalDetId,HGCSample>>&,
-                std::vector<std::pair<HGCalDetId, uint32_t > >&);
-        void linearize(const HGCalTriggerLightweightGeometryBase::geom_set& ,
-                const std::vector<HGCDataFrame<HGCalDetId,HGCSample>>&,
+        //void linearize(const HGCalTriggerGeometry::Module& ,
+                //const std::vector<HGCDataFrame<HGCalDetId,HGCSample>>&,
+                //std::vector<std::pair<HGCalDetId, uint32_t > >&);
+        void linearize(const std::vector<HGCDataFrame<HGCalDetId,HGCSample>>&,
                 std::vector<std::pair<HGCalDetId, uint32_t > >&);
 
-        void triggerCellSums(const HGCalTriggerGeometry::Module& ,
-                const std::vector<std::pair<HGCalDetId, uint32_t > >&,
-                data_type&);
-        void triggerCellSums(const HGCalTriggerLightweightGeometryBase& ,
+        //void triggerCellSums(const HGCalTriggerGeometry::Module& ,
+                //const std::vector<std::pair<HGCalDetId, uint32_t > >&,
+                //data_type&);
+        void triggerCellSums(const HGCalTriggerGeometryBase& ,
                 const std::vector<std::pair<HGCalDetId, uint32_t > >&,
                 data_type&);
         void bestChoiceSelect(data_type&);
