+++ L1Trigger/L1THGCal/plugins/fe_codecs/HGCalBestChoiceCodec.cc
@@ -17,55 +17,48 @@ HGCalBestChoiceCodec::HGCalBestChoiceCodec(const edm::ParameterSet& conf) : Code
 
 
 /*****************************************************************/
-void HGCalBestChoiceCodec::setDataPayloadImpl(const Module& mod, 
+void HGCalBestChoiceCodec::setDataPayloadImpl(const HGCalTriggerGeometryBase& geom, 
         const HGCEEDigiCollection& ee,
         const HGCHEDigiCollection& fh,
         const HGCHEDigiCollection& ) 
 /*****************************************************************/
 {
     data_.reset();
-    HGCalDetId moduleId(mod.moduleId());
     std::vector<HGCDataFrame<HGCalDetId,HGCSample>> dataframes;
     std::vector<std::pair<HGCalDetId, uint32_t > > linearized_dataframes;
-    // loop over EE or FH digis and fill digis belonging to that module
-    if(moduleId.subdetId()==ForwardSubdetector::HGCEE)
+    // convert ee and fh hit collections into the same object
+    if(ee.size()>0)
     {
         for(const auto& eedata : ee)
         {
-            if(mod.containsCell(eedata.id()))
+            dataframes.emplace_back(eedata.id());
+            for(int i=0; i<eedata.size(); i++)
             {
-                dataframes.emplace_back(eedata.id());
-                for(int i=0; i<eedata.size(); i++)
-                {
-                    dataframes.back().setSample(i, eedata.sample(i));
-                }
+                dataframes.back().setSample(i, eedata.sample(i));
             }
         }
     }
-    else if(moduleId.subdetId()==ForwardSubdetector::HGCHEF)
+    else if(fh.size()>0)
     {
         for(const auto& fhdata : fh)
         {
-            if(mod.containsCell(fhdata.id()))
+            dataframes.emplace_back(fhdata.id());
+            for(int i=0; i<fhdata.size(); i++)
             {
-                dataframes.emplace_back(fhdata.id());
-                for(int i=0; i<fhdata.size(); i++)
-                {
-                    dataframes.back().setSample(i, fhdata.sample(i));
-                }
+                dataframes.back().setSample(i, fhdata.sample(i));
             }
         }
     }
     // linearize input energy on 16 bits
-    codecImpl_.linearize(mod, dataframes, linearized_dataframes);
+    codecImpl_.linearize(dataframes, linearized_dataframes);
     // sum energy in trigger cells
-    codecImpl_.triggerCellSums(mod, linearized_dataframes, data_);
+    codecImpl_.triggerCellSums(geom, linearized_dataframes, data_);
     // choose best trigger cells in the module
     codecImpl_.bestChoiceSelect(data_);
 }
 
 /*****************************************************************/
-void HGCalBestChoiceCodec::setDataPayloadImpl(const Module& mod, 
+void HGCalBestChoiceCodec::setDataPayloadImpl(const HGCalTriggerGeometryBase& geom, 
         const l1t::HGCFETriggerDigi& digi)
 /*****************************************************************/
 {
