+++ L1Trigger/L1THGCal/src/fe_codecs/HGCalBestChoiceCodecImpl.cc
@@ -91,9 +91,34 @@ HGCalBestChoiceCodecImpl::data_type HGCalBestChoiceCodecImpl::decode(const std::
     return result;
 }
 
+//void HGCalBestChoiceCodecImpl::linearize(const HGCalTriggerGeometry::Module& mod,
+        //const std::vector<HGCDataFrame<HGCalDetId,HGCSample>>& dataframes,
+        //std::vector<std::pair<HGCalDetId, uint32_t > >& linearized_dataframes)
+//{
+    //double amplitude; uint32_t amplitude_int;
+   
+
+    //for(const auto& frame : dataframes) {//loop on DIGI
+        //for(const auto& tc_c : mod.triggerCellComponents()) { //treat only the HG cells in the considered module
+            //if(tc_c.second==frame.id()) { //treat if the DIGI detID is the same of the considered GC cell
+                //if (frame[2].mode()) {//TOT mode
+                    //amplitude =( floor(tdcOnsetfC_/adcLSB_) + 1.0 )* adcLSB_ + double(frame[2].data()) * tdcLSB_;
+                //}
+                //else {//ADC mode
+                    //amplitude = double(frame[2].data()) * adcLSB_;
+                //}
+
+                //amplitude_int = uint32_t (floor(amplitude/linLSB_+0.5));  
+                //if (amplitude_int>65535) amplitude_int = 65535;
+
+                //linearized_dataframes.push_back(std::make_pair (frame.id(), amplitude_int));
+            //}
+        //}
+    //}
+//}
+
 /*****************************************************************/
-void HGCalBestChoiceCodecImpl::linearize(const HGCalTriggerGeometry::Module& mod,
-        const std::vector<HGCDataFrame<HGCalDetId,HGCSample>>& dataframes,
+void HGCalBestChoiceCodecImpl::linearize(const std::vector<HGCDataFrame<HGCalDetId,HGCSample>>& dataframes,
         std::vector<std::pair<HGCalDetId, uint32_t > >& linearized_dataframes)
 /*****************************************************************/
 {
@@ -101,100 +126,66 @@ void HGCalBestChoiceCodecImpl::linearize(const HGCalTriggerGeometry::Module& mod
    
 
     for(const auto& frame : dataframes) {//loop on DIGI
-        for(const auto& tc_c : mod.triggerCellComponents()) { //treat only the HG cells in the considered module
-            if(tc_c.second==frame.id()) { //treat if the DIGI detID is the same of the considered GC cell
-                if (frame[2].mode()) {//TOT mode
-                    amplitude =( floor(tdcOnsetfC_/adcLSB_) + 1.0 )* adcLSB_ + double(frame[2].data()) * tdcLSB_;
-                }
-                else {//ADC mode
-                    amplitude = double(frame[2].data()) * adcLSB_;
-                }
-
-                amplitude_int = uint32_t (floor(amplitude/linLSB_+0.5));  
-                if (amplitude_int>65535) amplitude_int = 65535;
-
-                linearized_dataframes.push_back(std::make_pair (frame.id(), amplitude_int));
-            }
+        if (frame[2].mode()) {//TOT mode
+            amplitude =( floor(tdcOnsetfC_/adcLSB_) + 1.0 )* adcLSB_ + double(frame[2].data()) * tdcLSB_;
         }
-    }
-}
-
-/*****************************************************************/
-void HGCalBestChoiceCodecImpl::linearize(const HGCalTriggerLightweightGeometryBase::geom_set& cells_in_module,
-        const std::vector<HGCDataFrame<HGCalDetId,HGCSample>>& dataframes,
-        std::vector<std::pair<HGCalDetId, uint32_t > >& linearized_dataframes)
-/*****************************************************************/
-{
-    double amplitude; uint32_t amplitude_int;
-   
-
-    for(const auto& frame : dataframes) {//loop on DIGI
-        for(const auto& cell : cells_in_module) { //treat only the HG cells in the considered module
-            if(cell==frame.id()) { //treat if the DIGI detID is the same of the considered GC cell
-                if (frame[2].mode()) {//TOT mode
-                    amplitude =( floor(tdcOnsetfC_/adcLSB_) + 1.0 )* adcLSB_ + double(frame[2].data()) * tdcLSB_;
-                }
-                else {//ADC mode
-                    amplitude = double(frame[2].data()) * adcLSB_;
-                }
-
-                amplitude_int = uint32_t (floor(amplitude/linLSB_+0.5));  
-                if (amplitude_int>65535) amplitude_int = 65535;
-
-                linearized_dataframes.push_back(std::make_pair (frame.id(), amplitude_int));
-            }
+        else {//ADC mode
+            amplitude = double(frame[2].data()) * adcLSB_;
         }
+
+        amplitude_int = uint32_t (floor(amplitude/linLSB_+0.5));  
+        if (amplitude_int>65535) amplitude_int = 65535;
+
+        linearized_dataframes.push_back(std::make_pair (frame.id(), amplitude_int));
     }
 }
   
 
-/*****************************************************************/
-void HGCalBestChoiceCodecImpl::triggerCellSums(const HGCalTriggerGeometry::Module& mod,  const std::vector<std::pair<HGCalDetId, uint32_t > >& linearized_dataframes, data_type& data)
-/*****************************************************************/
-{
-    std::map<HGCalDetId, uint32_t> payload;
-    // sum energies in trigger cells
-    for(const auto& frame : linearized_dataframes)
-    {
-        // FIXME: only EE
-        HGCalDetId cellid(frame.first);
-        // find trigger cell associated to cell
-        uint32_t tcid(0);
-        for(const auto& tc_c : mod.triggerCellComponents())
-        {
-            if(tc_c.second==cellid)
-            {
-                tcid = tc_c.first;
-                break;
-            }
-        }
-        if(!tcid)
-        {
-            throw cms::Exception("BadGeometry")
-                << "Cannot find trigger cell corresponding to HGC cell "<<cellid<<"\n";
-        }
-        HGCalDetId triggercellid( tcid );
-        payload.insert( std::make_pair(triggercellid, 0) ); // do nothing if key exists already
-        // FIXME: need to transform ADC and TDC to the same linear scale on 12 bits
-        uint32_t value = frame.second; // 'value' has to be a 12 bit word
-        payload[triggercellid] += value; // 32 bits integer should be largely enough (maximum 7 12-bits sums are done)
+//void HGCalBestChoiceCodecImpl::triggerCellSums(const HGCalTriggerGeometry::Module& mod,  const std::vector<std::pair<HGCalDetId, uint32_t > >& linearized_dataframes, data_type& data)
+//{
+    //std::map<HGCalDetId, uint32_t> payload;
+    //// sum energies in trigger cells
+    //for(const auto& frame : linearized_dataframes)
+    //{
+        //// FIXME: only EE
+        //HGCalDetId cellid(frame.first);
+        //// find trigger cell associated to cell
+        //uint32_t tcid(0);
+        //for(const auto& tc_c : mod.triggerCellComponents())
+        //{
+            //if(tc_c.second==cellid)
+            //{
+                //tcid = tc_c.first;
+                //break;
+            //}
+        //}
+        //if(!tcid)
+        //{
+            //throw cms::Exception("BadGeometry")
+                //<< "Cannot find trigger cell corresponding to HGC cell "<<cellid<<"\n";
+        //}
+        //HGCalDetId triggercellid( tcid );
+        //payload.insert( std::make_pair(triggercellid, 0) ); // do nothing if key exists already
+        //// FIXME: need to transform ADC and TDC to the same linear scale on 12 bits
+        //uint32_t value = frame.second; // 'value' has to be a 12 bit word
+        //payload[triggercellid] += value; // 32 bits integer should be largely enough (maximum 7 12-bits sums are done)
 
-    }
-    // fill data payload
-    for(const auto& id_value : payload)
-    {
-        uint32_t id = id_value.first.cell();
-        if(id>=nCellsInModule_) 
-        {
-            throw cms::Exception("BadGeometry")
-                << "Number of trigger cells in module too large for available data payload\n";
-        }
-        data.payload.at(id) = id_value.second;
-    }
-}
+    //}
+    //// fill data payload
+    //for(const auto& id_value : payload)
+    //{
+        //uint32_t id = id_value.first.cell();
+        //if(id>=nCellsInModule_) 
+        //{
+            //throw cms::Exception("BadGeometry")
+                //<< "Number of trigger cells in module too large for available data payload\n";
+        //}
+        //data.payload.at(id) = id_value.second;
+    //}
+//}
 
 /*****************************************************************/
-void HGCalBestChoiceCodecImpl::triggerCellSums(const HGCalTriggerLightweightGeometryBase& geometry,  const std::vector<std::pair<HGCalDetId, uint32_t > >& linearized_dataframes, data_type& data)
+void HGCalBestChoiceCodecImpl::triggerCellSums(const HGCalTriggerGeometryBase& geometry,  const std::vector<std::pair<HGCalDetId, uint32_t > >& linearized_dataframes, data_type& data)
 /*****************************************************************/
 {
     std::map<HGCTriggerHexDetId, uint32_t> payload;
