+++ L1Trigger/L1THGCal/test/BuildFile.xml
@@ -11,7 +11,7 @@
 <use   name="DataFormats/L1Trigger"/>
 <use   name="SimDataFormats/CaloTest"/>
 
-<library name="testL1TriggerL1THGCal"  file="HGCalTriggerGeomTester.cc,HGCalTriggerBestChoiceTester.cc,HGCalTriggerLightweightGeomTester.cc,HGCalTriggerBestChoiceMonitor.cc">
+<library name="testL1TriggerL1THGCal"  file="HGCalTriggerBestChoiceTester.cc,HGCalTriggerGeomTester.cc,HGCalTriggerBestChoiceMonitor.cc">
 <flags   EDM_PLUGIN="1"/>
 </library>
